Added EntangleRegistry for CM and their copies.
So, I decided to keep entanglement. This requires keeping track of
machines and their entangled clones. They are being registered in
the EntangleRegistry which can be queried for remaining machines
using the same room. With this machines can be indexed and this
additional index can be used for another abstraction inside the
AbstractSharedStorages.
This can in AE2s can be used to have a different IGridNode for each
instance of the entangled CM.

Fixes #68