Create pt_PT Language
This file was created to help  thraaawn adding more languages to his mod. 
Even though this is Portuguese from Portugal, Brazilian Portuguese can be just a copy of this file. There would be some differencies, but it would at least be a way for people who are not used to english to understand.