Added crafting mechanics using miniaturization fluid
The simplest example would be placing a gold block down and pouring
a bucket of miniaturization fluid on top of it so the top and sides
are covered with fluid. Then throw one piece of redstone dust into
the liquid on top of the block. The gold should disappear, the
liquid drain and 6 pieces of tunnel tools appear.

Iron in a 3x1x3 shape with 4 buckets of fluid on top gives 64 wall
blocks. Creating a hollow 5x5x5 (or 7, 9, 11, 13, 15) cube with them
and pouring 9 buckets of fluid over it so everything is covered, gives
the corresponding compact machine.

I might switch some of the source materials for others.