Capability Null Handlers
EnderIO is fine with server-only capabilities, xnet on the other hand
checks on the clientside whether a connector needs to render a connection,
which is absolutely valid.

We do not know on the clientside whether a machine supports a certain
capability on a specific side, because we don't have access to the
machineworld on the client if he is not located in that world.
And because we don't want to say "yes", we have that capability and then
return null when actually asked for it, we have to fake the capabilities
with implementations that do nothing, e.g. they neither accept items/
fluid/energy nor do they provide them.

On the server we only do that optionally: if the CM does not provide the
needed capability on that side. Otherwise and by default we still pass
along all capabilities - it might just be that the client does not know
about it, which is usually only a visual problem anyway.
But this allows machines to keep the configuration for CM blocks when
their capabilities suddenly change, e.g. because the player broke a chest
next to a tunnel.