Off-hand use for the personal shrinking device
Requires the PSD to be filled with miniaturization fluid.
Applies the potion effect every few ticks and drains a bit of the
fluid. One millibucket is one tick of being shrinked, so a fully
charged PSD holding 4000mb of fluid lasts for 3 minutes and 20
seconds.

I should probably make this a bit cheaper?