Reworked crafting system, json based
Note: this commit changes the default location of the config file!
If you've changed any settings, copy your old file over to the new
location!

This also required a reworked multiblock detection and jei rendering
procedure. It also makes it a lot harder to calculate the amount of
required buckets. This is just an estimate for now and should be good
enough.
There is some mischief people can do with custom recipes, which can
in the end lead to server crashes because of malformed json or invalid
recipe structure. I'll try and catch them in the future and give some
more feedback to the user.

Just remember that each recipe layer has to have the same size!