Fixed recipe loading and extracting
This needed special treatment depending on whether you run in a dev
or productive environment, i.e. are the included recipes in a jar or
on the filesystem.
With this commit recipes do not need to be hardcoded anymore to allow
being extracted/loaded and it is reported where a recipe got loaded
from during init.