Removed cruft I don't want to keep + fixed deprecations
Basically all of this comes from my decision to strip the mod down to its
core essentials so that I have a realistic chance of maintaining it in the
future. I simply do not have the time for it anymore. Hence, this commit
makes quite a large cut to some of the newer features:
- Removed everything miniaturization fluid related, i.e. the fluid and
  the mini fluid drops. The algorithm to determine whether everything
  is covered in fluid is simply too complicated.
- Removed player shrinking feature completely. This is just too difficult
  to maintain compatibility with other mods and would require frequent
  fixes and updates.

But: the crafting mechanism will kind of stay, just not with fluid but
with a simple frame you'd have to build around it. At least thats the
plan.

All of this speeds up the development of 1.12 version a lot. Actually
in creative mode this might be quite stable at this point in time.