Readded crafting mechanic via Miniaturization Field Projector
Since there is no miniaturization fluid anymore we needed a different way
to perform the miniaturization crafting. We are now doing this by spanning
a miniaturization field between four Miniaturization Field Projectors
facing each other.

We are currently representing the field with a lot of particles. A LOT.
Not sure every PC can handle this, we'll probably either need an
alternative or at least a config option to tune down the amount of
particles generated.

We've also borrowed the Sensor Model from OCSensors for the projector. It
fits the mods theme nicely with just a small model tweak + retexture.