Gui for Machines: show 3D view of the machine interior
First of all a bit of history: i've always wanted to do this. The Looking
Glass Mod by xCompWiz helped me a lot in the 1.7.10 days when I literally
had no idea about minecrafts rendering system and could just use that to
render other worlds on the client side.

Today a big chunk of what was needed back then is a lot easier to do in
recent minecraft versions, which allowed me to implement at least parts of
this feature natively without other mod dependencies. Whats missing and
what will never happen is the rendering of TileEntitySpecialRenderers.
They simply require too much maintenance because they need a fully fledged
fake world. Simply rendering blocks can be done using a simple
IBlockAccess implementation on top of a single chunk, which is exactly
what we are doing now.

We've borrowed some parts of the AnvilChunkLoader class to be able to send
the machines machineworld-chunk as nbt over the wire. At the moment this
is done once when the client opens the GUI, but we'll either add a refresh
button or do this every few ticks automatically.