Reworked recipes to be less grindy
These were still the recipes I've added during development for
testing and they were never meant to be released that way. Oops.

The amount of iron needed was ridiculous. The weird way to get
the Personal Shrinking Device using Miniaturization Crafting was
removed - you'll need the PSD to learn how to do it in the first
place.