Improve retrieval of tunnels after cycling through them
First try to insert them into the players inventory. If that fails
spawn them at the players location, not in front of the actual wall
block. That position might have a transfer node or chest or something
in front of it making the tunnel item shoot off in a random direction.