Render TESRs in the Machine view GUI
This requires a ProxyWorld to provide some additional data, which is
mostly grabbed from the clients current world, so it might not be 100%
accurate, e.g. when weather or daytime influences rendering.
It also might not work at all and crash. I hope it doesn't. Requires a
bit more testing first.

Also improved block rendering by getting the extended blockstate, which
some blocks require to e.g. rotate properly.

TESRs can not be rendered using display lists though, so this obviously
causes a bit of a hit in performance, but not more than being in the
machine itself would be.

All in all this is a major improvement and we're getting closer to
rendering the blocks content in-world in Compact Machines 4 ;)