Added extra tile data providers for the proxy world
We can not send update packets to the client as the client would
process the packets on the world he's currently in and not in our
fake proxy world. We'd have to patch the nethandler etc and we don't
want to do that.

What we're doing instead is definitely a hack. We basically provide
a way to add extra tile entity data to the chunk being sent to the
client.
We can use reflection to get the data on the server side and set it
on the client side. All that needs to be done is extending and
annotating the AbstractExtraTileDataProvider class. With a bit of
reflection the necessary data can be transferred from the server
to the client.

This commit already adds a provider for Cofh TileGrid tile entities
and their connection state, i.e. ducts are now being rendered
properly.