Gui rework to use own widget based framework
Uffz, but this is bigger than it seems as it basically just cleans up the GuiMachine classes into separate widgets. It allows for much easier reasoning about the GUI and makes it a lot easier to add new functionality.
But also touches some parts that might be crashy in regards to ConcurrentModificationExceptions. It should not happen, but I can't fully guarantee that. Needs a bit more testing.

This PR also includes a fix for #388 and improvements for the compact skies world, where players are now teleported to their hub machine when returning to the overworld (this only works for newly claimed machines in the hub).
New features are the ability to rename machines in the machine GUI and easier adding/removing of whitelist entries.