Added basic API to allow access to connected blocks
Someone in some issue somewhere said there is no support for a given
connection type through tunnels, because compact machines does not have
an API and only allows Capability based connections.

This commit changes that. Other mods can now ask Machines and Tunnels
to which WorldServer and BlockPos they are connected to. Maybe this
results in more possible connection types and less spam in the issue
inbox.