Add 2-dimensional grid support to machine generation. Fixes #429
This change was designed with backwards compatability in mind, but is not API compatible.
This means the version has been bumped to 3.1.0 in line with semantic versioning.
This new version **MUST** be the same on both client and server. It *might* be possible to not have this on the client, but it is unsupported.

Summary of changes:
Add 2d grid generation
Support older 3.0.18 file formats
Update build.gradle
Manually include RS API
Add /compactmachines3 documentation command
Minor bug fixes
Code formatting
Translation overhaul
Logger refactoring
and other stuff I have missed